backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx[http2]>=0.26.0,<0.29.0
pydantic>=2.11.7,<3.0.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv
groq
